backend/requirements.txt
flask==3.0.3
flask-cors==4.0.1
flask-jwt-extended==4.6.0
flask-sqlalchemy==3.1.1
flask-bcrypt==1.0.1
flask-limiter==3.7.0
python-dotenv==1.0.1
groq==0.9.0
langchain==0.2.14
langchain-groq==0.1.9
transformers==4.44.2
torch==2.4.0
pillow==10.4.0
sentence-transformers==3.0.1
# openai-whisper==20231117
pydantic==2.8.2
marshmallow==3.21.3
secure-smtplib==0.1.1